config: rk3066: mk808: move config to new rk3066 bootflow
We are moving to new rk3066 bootflow.

To flash firmware to the board we boot TPL from usb, next
we boot SPL with rockusb and nand support enabled.

To boot u-boot we use TPL to init ram and clks and then
we boot U-BOOT proper. All is loaded by bootrom

Signed-off-by: Paweł Jarosz <paweljarosz3691@gmail.com>