mtd: nand: rockchip: add minimal spl nand driver
To upload the rk3066 firmware, the SPL must fit within 64kb
(-2 bytes per checksum). Using full nand rockchip driver causes by
dependencies that spl size exceeds this limit. In order to fit in this
dependency we add a minimal spl driver which is only used to write the
bootloader to the beginning of the nand memory and nothing else.

By finding a loophole in bootloader reading by bootrom we were able to
double the size allocated to u-boot by writing (u32) "0" to the first
four bytes of oob data. Previously, this limit was 256KB (-2048 bytes).
Now it is 512KB and it is enough to write u-boot proper + tpl to one
erase block.

The driver mimics bootrom in writing the bootloader. It does not use the
full size of the nand page. It uses only 2KB in each page. Which, with,
say, a 16KB page and a 4MB erase block, gives 256 2KB pages for u-boot.

The driver mimics rkusbplug from rockchip by writing a copy of the
bootloader to all bootrom blocks.

Signed-off-by: Paweł Jarosz <paweljarosz3691@gmail.com>