rockusb: add spl nand read, write, erase support
We now use the minimal spl nand driver to write the bootloader
to the nand memory. So let's add support for this write in the
rockusb driver.

To maintain support for the upgrade_tool, rkdeveloptool write is
done to 0x40 lba (which is 512B in size). This way we distinguish
if the client is trying to write the bootloader or something else.

In spl we only allow the bootloader to be written. Writing the system
to nand will be done by u-boot proper nand driver.

We use rksd mkimage format to write and spread 2KB chunks of bootloader
to every page in erase block.

Signed-off-by: Paweł Jarosz <paweljarosz3691@gmail.com>